add show in picker exp + refactor (#1723)
* add show in picker exp + refactor

* more concise

* Update src/platform/endpoint/node/modelMetadataFetcher.ts


* remove unnecessary async

* no really remove it

---------
