+++ src/platform/endpoint/node/modelMetadataFetcher.ts
@@ -12,7 +12,7 @@ import { Disposable } from '../../../util/vs/base/common/lifecycle';
 import { generateUuid } from '../../../util/vs/base/common/uuid';
 import { IInstantiationService, ServicesAccessor } from '../../../util/vs/platform/instantiation/common/instantiation';
 import { IAuthenticationService } from '../../authentication/common/authentication';
-import { IConfigurationService } from '../../configuration/common/configurationService';
+import { ConfigKey, IConfigurationService } from '../../configuration/common/configurationService';
 import { IEnvService } from '../../env/common/envService';
 import { ILogService } from '../../log/common/logService';
 import { IFetcherService } from '../../networking/common/fetcherService';
@@ -23,7 +23,6 @@ import { ITelemetryService } from '../../telemetry/common/telemetry';
 import { ICAPIClientService } from '../common/capiClient';
 import { ChatEndpointFamily, IChatModelInformation, ICompletionModelInformation, IEmbeddingModelInformation, IModelAPIResponse, isChatModelInformation, isCompletionModelInformation, isEmbeddingModelInformation } from '../common/endpointProvider';
 import { ModelAliasRegistry } from '../common/modelAliasRegistry';
-import { getMaxPromptTokens } from './chatEndpoint';
 
 export interface IModelMetadataFetcher {
 
@@ -137,19 +136,25 @@ export class ModelMetadataFetcher extends Disposable implements IModelMetadataFe
 	 * @returns The resolved model with proper exp overrides and token counts
 	 */
 	private async _hydrateResolvedModel(resolvedModel: IModelAPIResponse | undefined): Promise<IModelAPIResponse> {
-		resolvedModel = resolvedModel ? await this._findExpOverride(resolvedModel) : undefined;
+		resolvedModel = resolvedModel ? await this._getExpModelOverride(resolvedModel) : undefined;
 		if (!resolvedModel) {
 			throw this._lastFetchError;
 		}
 
 		// If it's a chat model, update max prompt tokens based on settings + exp
 		if (isChatModelInformation(resolvedModel) && (resolvedModel.capabilities.limits)) {
-			resolvedModel.capabilities.limits.max_prompt_tokens = getMaxPromptTokens(this._configService, this._expService, resolvedModel);
+			resolvedModel.capabilities.limits.max_prompt_tokens = this._getMaxPromptTokensOverride(resolvedModel);
 			// Also ensure prompt tokens + output tokens <= context window. Output tokens is capped to max 15% input tokens
 			const outputTokens = Math.floor(Math.min(resolvedModel.capabilities.limits.max_output_tokens ?? 4096, resolvedModel.capabilities.limits.max_prompt_tokens * 0.15));
 			const contextWindow = resolvedModel.capabilities.limits.max_context_window_tokens ?? (outputTokens + resolvedModel.capabilities.limits.max_prompt_tokens);
 			resolvedModel.capabilities.limits.max_prompt_tokens = Math.min(resolvedModel.capabilities.limits.max_prompt_tokens, contextWindow - outputTokens);
 		}
+
+		// If it's a chat model, update showInModelPicker based on experiment overrides
+		if (isChatModelInformation(resolvedModel)) {
+			resolvedModel.model_picker_enabled = this._getShowInModelPickerOverride(resolvedModel);
+		}
+
 		if (resolvedModel.preview && !resolvedModel.name.endsWith('(Preview)')) {
 			// If the model is a preview model, we append (Preview) to the name
 			resolvedModel.name = `${resolvedModel.name} (Preview)`;
@@ -335,7 +340,7 @@ export class ModelMetadataFetcher extends Disposable implements IModelMetadataFe
 		}
 	}
 
-	private async _findExpOverride(resolvedModel: IModelAPIResponse): Promise<IModelAPIResponse | undefined> {
+	private async _getExpModelOverride(resolvedModel: IModelAPIResponse): Promise<IModelAPIResponse | undefined> {
 		// This is a mapping of model id to model id. Allowing us to override the request for any model with a different model
 		let modelExpOverrides: { [key: string]: string } = {};
 		const expResult = this._expService.getTreatmentVariable<string>('copilotchat.modelOverrides');
@@ -358,6 +363,57 @@ export class ModelMetadataFetcher extends Disposable implements IModelMetadataFe
 			resolvedModel = experimentalModel ?? resolvedModel;
 		}
 		return resolvedModel;
+	}
+
+	// get ChatMaxNumTokens from config for experimentation
+	private _getMaxPromptTokensOverride(chatModelInfo: IChatModelInformation): number {
+		// check debug override ChatMaxTokenNum
+		const chatMaxTokenNumOverride = this._configService.getConfig(ConfigKey.Internal.DebugOverrideChatMaxTokenNum); // can only be set by internal users
+		// Base 3 tokens for each OpenAI completion
+		let modelLimit = -3;
+		// if option is set, takes precedence over any other logic
+		if (chatMaxTokenNumOverride > 0) {
+			modelLimit += chatMaxTokenNumOverride;
+			return modelLimit;
+		}
+
+		let experimentalOverrides: Record<string, number> = {};
+		try {
+			const expValue = this._expService.getTreatmentVariable<string>('copilotchat.contextWindows');
+			experimentalOverrides = JSON.parse(expValue ?? '{}');
+		} catch {
+			// If the experiment service either is not available or returns a bad value we ignore the overrides
+		}
+
+		// If there's an experiment that takes precedence over what comes back from CAPI
+		if (experimentalOverrides[chatModelInfo.id]) {
+			modelLimit += experimentalOverrides[chatModelInfo.id];
+			return modelLimit;
+		}
+
+		// Check if CAPI has prompt token limits and return those
+		if (chatModelInfo.capabilities?.limits?.max_prompt_tokens) {
+			modelLimit += chatModelInfo.capabilities.limits.max_prompt_tokens;
+			return modelLimit;
+		} else if (chatModelInfo.capabilities.limits?.max_context_window_tokens) {
+			// Otherwise return the context window as the prompt tokens for cases where CAPI doesn't configure the prompt tokens
+			modelLimit += chatModelInfo.capabilities.limits.max_context_window_tokens;
+			return modelLimit;
+		}
+
+		return modelLimit;
+	}
+
+	private _getShowInModelPickerOverride(resolvedModel: IModelAPIResponse): boolean {
+		let modelPickerOverrides: Record<string, boolean> = {};
+		const expResult = this._expService.getTreatmentVariable<string>('copilotchat.showInModelPicker');
+		try {
+			modelPickerOverrides = JSON.parse(expResult || '{}');
+		} catch {
+			// No-op if parsing experiment fails
+		}
+
+		return modelPickerOverrides[resolvedModel.id] ?? resolvedModel.model_picker_enabled;
 	}
 }
